Extract to its own file, add tests
